Refactor: MemberPage 컴포넌트 구조 분리 및 fetch 로직 개선
