+++ app/(kahlua)/admin/member/page.tsx
@@ -1,119 +1,29 @@
 'use client';
 
-import { useState, useEffect, useMemo } from 'react';
-import { authInstance } from '@/api/auth/axios';
+import { useState } from 'react';
 import Header from '@/components/admin/Header';
 import MemberStatusIcon from '@/components/admin/member/MemberStatusIcon';
+import TableSection from '@/components/admin/member/TableSection';
 import WaitingIcon from '@/public/image/admin/WaitingIcon.svg';
 import WaitingMobileIcon from '@/public/image/admin/WaitingMobileIcon.svg';
 import CompletedIcon from '@/public/image/admin/CompletedIcon.svg';
 import CompletedMobileIcon from '@/public/image/admin/CompletedMobileIcon.svg';
-import SearchBar from '@/components/admin/member/SearchBar';
-import MemberTable from '@/components/admin/member/MemberTable';
 
 const MemberPage = () => {
   const [isWaiting, setIsWaiting] = useState(false);
-  const [searchQuery, setSearchQuery] = useState('');
-
-  const [members, setMembers] = useState<any[]>([]);
-  const [allMembers, setAllMembers] = useState<any[]>([]);
-  const [totalPages, setTotalPages] = useState(0);
   const [currentPage, setCurrentPage] = useState(0);
   const [waitingCount, setWaitingCount] = useState(0);
   const [completedCount, setCompletedCount] = useState(0);
-  const [isSearching, setIsSearching] = useState(false);
 
-  const [isMounted, setIsMounted] = useState(false);
-
-  const pageSize = 8;
-
-  const fetchPageMembers = async () => {
-    try {
-      const response = await authInstance.get('/admin/users', {
-        params: {
-          approvalFilter: isWaiting ? 'PENDING' : 'ALL',
-          page: currentPage,
-          size: pageSize,
-        },
-      });
-      const { content, pageInfo, pendingCount, approvedCount } =
-        response.data.result;
-      setMembers(content);
-      setTotalPages(pageInfo.totalPages);
-      setWaitingCount(pendingCount);
-      setCompletedCount(approvedCount);
-    } catch (error) {
-      console.error('멤버 데이터 불러오기 실패:', error);
-    }
-  };
-
-  const fetchAllMembers = async () => {
-    try {
-      const response = await authInstance.get('/admin/users', {
-        params: {
-          approvalFilter: isWaiting ? 'PENDING' : 'ALL',
-          page: 0,
-          size: 9999,
-        },
-      });
-      const { content } = response.data.result;
-      setAllMembers(content);
-    } catch (error) {
-      console.error('전체 멤버 데이터 불러오기 실패:', error);
-    }
-  };
-
-  useEffect(() => {
-    if (!isSearching && searchQuery === '') {
-      fetchPageMembers();
-    }
-  }, [currentPage, isWaiting]);
-
-  useEffect(() => {
-    if (searchQuery.length > 0) {
-      if (!isSearching) {
-        fetchAllMembers();
-        setIsSearching(true);
-        setCurrentPage(0);
-      }
-    } else {
-      if (isSearching) {
-        setIsSearching(false);
-        setCurrentPage(0);
-      }
-    }
-  }, [searchQuery]);
-
-  const handleSearchChange = (value: string) => {
-    setSearchQuery(value);
+  const handleToggleWaiting = () => {
+    setIsWaiting((prev) => !prev);
     setCurrentPage(0);
   };
-
-  const displayedMembers = useMemo(() => {
-    if (isSearching) {
-      const filtered = allMembers.filter((member) =>
-        member.name.includes(searchQuery)
-      );
-      const start = currentPage * pageSize;
-      const end = start + pageSize;
-      return filtered.slice(start, end);
-    } else {
-      return members;
-    }
-  }, [isSearching, searchQuery, allMembers, members, currentPage]);
-
-  useEffect(() => {
-    if (isSearching) {
-      const filtered = allMembers.filter((member) =>
-        member.name.includes(searchQuery)
-      );
-      setTotalPages(Math.ceil(filtered.length / pageSize));
-    }
-  }, [searchQuery, allMembers]);
 
   return (
     <div className="w-full h-auto min-h-[calc(100vh-390px)] flex flex-col mt-16 text-black font-pretendard items-center">
       <Header subtitle="깔루아 멤버 정보 관리" />
+
       <div className="w-full max-pad:px-4">
         <div className="flex flex-col h-auto min-[1500px]:w-[1200px] min-[834px]:w-[786px] max-pad:max-w-[500px] mt-[134px] max-dt:mt-[107px] max-pad:mt-6 mx-auto">
           <div className="flex max-pad:flex-col gap-[75px] max-pad:gap-[10px]">
@@ -133,33 +43,16 @@ const MemberPage = () => {
             />
           </div>
 
-          <div className="flex self-end max-dt:mt-10 max-pad:mt-6">
-            <SearchBar
-              searchQuery={searchQuery}
-              setSearchQuery={handleSearchChange}
-            />
-          </div>
-
-          <div className="w-full mt-5 max-dt:mt-[35px] max-pad:mt-4">
-            <MemberTable
-              isWaiting={isWaiting}
-              searchQuery={searchQuery}
-              members={displayedMembers}
-              setMembers={setMembers}
-              currentPage={currentPage}
-              totalPages={totalPages}
-              onPageChange={setCurrentPage}
-            />
-          </div>
+          <TableSection
+            isWaiting={isWaiting}
+            setWaitingCount={setWaitingCount}
+            setCompletedCount={setCompletedCount}
+            currentPage={currentPage}
+            setCurrentPage={setCurrentPage}
+          />
 
           <span
-            onClick={() => {
-              setIsWaiting((prev) => !prev);
-              setCurrentPage(0);
-              setSearchQuery('');
-              setIsSearching(false);
-              setAllMembers([]);
-            }}
+            onClick={handleToggleWaiting}
             className="flex self-end mt-[-32px] max-pad:mt-[-20px] text-2xl max-dt:text-[20px] max-pad:text-sm font-semibold cursor-pointer"
           >
             {isWaiting ? '전체 보기' : '승인 대기만 보기'}
